move node.rs to new cluster module
