+++ scylla/src/cluster/locator/test.rs
@@ -4,10 +4,9 @@ use uuid::Uuid;
 
 use super::{ReplicaLocator, ReplicaSet};
 use crate::cluster::metadata::{Keyspace, Metadata, Peer, Strategy};
+use crate::cluster::{Node, NodeAddr, NodeRef};
 use crate::connection::PoolConfig;
 use crate::sharding::Token;
-use crate::transport::Node;
-use crate::transport::{NodeAddr, NodeRef};
 
 use std::collections::HashSet;
 use std::sync::Arc;
