+++ scylla/src/cluster/mod.rs
@@ -1,6 +1,13 @@
 mod cluster;
 pub mod locator;
 pub mod metadata;
+mod node;
 
 pub(crate) use cluster::{Cluster, ClusterNeatDebug};
 pub use cluster::{ClusterData, Datacenter};
+
+pub(crate) use node::{resolve_contact_points, resolve_hostname, ResolvedContactPoint};
+pub use node::{KnownNode, Node, NodeAddr, NodeRef};
+
+#[cfg(feature = "cloud")]
+pub(crate) use node::CloudEndpoint;
