+++ scylla/src/connection/connection.rs
@@ -42,9 +42,9 @@ use std::{
 #[cfg(feature = "cloud")]
 use crate::cloud::CloudConfig;
 use crate::cluster::metadata::{PeerEndpoint, UntranslatedEndpoint, UntranslatedPeer};
+use crate::cluster::NodeAddr;
 use crate::transport::errors::{BadKeyspaceName, DbError, QueryError};
 use crate::transport::iterator::RowIterator;
-use crate::transport::NodeAddr;
 
 use crate::batch::{Batch, BatchStatement};
 use crate::frame::protocol_features::ProtocolFeatures;
@@ -1863,8 +1863,8 @@ mod tests {
     use super::open_connection;
     use super::ConnectionConfig;
     use crate::cluster::metadata::UntranslatedEndpoint;
+    use crate::cluster::ResolvedContactPoint;
     use crate::query::Query;
-    use crate::transport::node::ResolvedContactPoint;
     use crate::utils::test_utils::unique_keyspace_name;
     use crate::{IntoTypedRows, SessionBuilder};
     use futures::{StreamExt, TryStreamExt};
