+++ scylla/src/transport/mod.rs
@@ -4,7 +4,6 @@ pub mod host_filter;
 pub mod iterator;
 pub mod load_balancing;
 pub(crate) mod metrics;
-pub(crate) mod node;
 pub mod partitioner;
 pub mod query_result;
 pub mod retry_policy;
@@ -23,5 +22,3 @@ mod cql_collections_test;
 mod cql_types_test;
 #[cfg(test)]
 mod cql_value_test;
-
-pub use node::{KnownNode, Node, NodeAddr, NodeRef};
