+++ scylla/tests/integration/execution_profiles.rs
@@ -5,9 +5,9 @@ use crate::utils::test_with_3_node_cluster;
 use assert_matches::assert_matches;
 use scylla::batch::BatchStatement;
 use scylla::batch::{Batch, BatchType};
+use scylla::cluster::NodeRef;
 use scylla::query::Query;
 use scylla::statement::SerialConsistency;
-use scylla::transport::NodeRef;
 use scylla::{
     cluster::ClusterData,
     load_balancing::{LoadBalancingPolicy, RoutingInfo},
